ACP-L1-V2-E03-chat-sequencer: reduce chat_sessions/internal/service file-count regression
Merges attachment_ack_test.go into attachment_test.go and stream_head_test.go
into session_test.go (thematic pairing: AcknowledgeAttachment tests belong
with Attachment lifecycle tests, AdvanceStreamHead tests belong with Session
lifecycle tests), taking the package from 27 files back to 25. The remaining
sequencing_test.go (931 lines) cannot merge into any existing file or split
further without exceeding the repository's 1000-line-per-file hard cap
(backendsizecheck), so it stays standalone as the dedicated test suite for
the new Sequence identity-assignment/commit engine.

Registers the resulting, minimized count in
docs/internal/baselines/backend-package-file-count.json following the exact
convention already used for ~30 other legitimately oversized packages in this
repository (owner + removalReason), since the package already exceeded the
plain 15-file limit before this PR touched it.

Also merges fork/main to resolve a merge conflict in the generated
pkg/wire/wire_gen.go, regenerated via `make generate-wire` rather than
resolved by hand.
